refactor: change file names, delete unnecessary codes, fixed #31
