--- src/components/dashboard/common/PDFViewerNoSSR.tsx
@@ -1,9 +1,0 @@
-import dynamic from "next/dynamic";
-
-const NoSSR = dynamic(() => import("./PdfViewer"), {
-  ssr: false,
-});
-
-export default function PDFViewerWithNoSSR() {
-  return <NoSSR />;
-}
