+++ src/components/dashboard/report/organisms/ReportViewer.tsx
@@ -1,10 +1,10 @@
-import PDFViewerWithNoSSR from "../../common/PDFViewerNoSSR";
+import PDFViewer from "@/components/dashboard/common/PDFViewer";
 
 export default function ReportViewer() {
   return (
     <div className="flex flex-col gap-3 p-8 rounded-xl bg-gray-50">
       <h3 className="font-semibold text-gray-900">{"작성한 보고서"}</h3>
-      <PDFViewerWithNoSSR />
+      <PDFViewer />
     </div>
   );
 }
